requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai>=2.0.0
google-generativeai>=0.8.3
yfinance
python-dotenv